update on commercial paper

+++ FintrakBanking.Repositories/Credit/AppraisalMemorandumRepository.cs
@@ -658,7 +658,7 @@ namespace FintrakBanking.Repositories.Credit
                             loanPreliminaryEvaluationId = x.a.LOANPRELIMINARYEVALUATIONID,
                             //customerName = x.a.CustomerId.HasValue ? x.a.tbl_Customer.FirstName + " " + x.a.tbl_Customer.MiddleName + " " + x.a.tbl_Customer.LastName : "",
                             operationId = x.a.OPERATIONID,
-                            productClassId = x.a.PRODUCTCLASSID,
+                            productClassId = (short)x.a.PRODUCTCLASSID,
                         })
                         .GroupBy(d => d.loanApplicationId)
                         .Select(g => g.OrderByDescending(b => b.approvalTrailId).FirstOrDefault())
@@ -746,7 +746,7 @@ namespace FintrakBanking.Repositories.Credit
                 loanPreliminaryEvaluationId = x.a.LOANPRELIMINARYEVALUATIONID,
                 //customerName = x.a.CustomerId.HasValue ? x.a.tbl_Customer.FirstName + " " + x.a.tbl_Customer.MiddleName + " " + x.a.tbl_Customer.LastName : "",
                 operationId = x.a.OPERATIONID,
-                productClassId = x.a.PRODUCTCLASSID,
+                productClassId = (short)x.a.PRODUCTCLASSID,
             })
             .OrderByDescending(x => x.applicationDate)
             .ThenByDescending(x => x.loanApplicationId)
@@ -997,7 +997,7 @@ namespace FintrakBanking.Repositories.Credit
                             loanPreliminaryEvaluationId = x.a.LOANPRELIMINARYEVALUATIONID,
                             //customerName = x.a.CustomerId.HasValue ? x.a.tbl_Customer.FirstName + " " + x.a.tbl_Customer.MiddleName + " " + x.a.tbl_Customer.LastName : "",
                             operationId = x.a.OPERATIONID,
-                            productClassId = x.a.PRODUCTCLASSID,
+                            productClassId = (short)x.a.PRODUCTCLASSID,
                         })
                         .GroupBy(d => d.loanApplicationId)
                         .Select(g => g.OrderByDescending(b => b.approvalTrailId).FirstOrDefault())
